+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 1.0.0 (unreleased)
 ------------------
 
+- #5 Allow to store the raw ASTM messages
 - #4 Improve ASTM Protocol for Multi connections
 - #3 Fix ASTM message splitting and added tests
 - #1 Ensure output messages are LIS2-A2 compliant